chore: 커뮤니티 구조 변경
