+++ backend/src/main/java/com/ourdressingtable/community/comment/domain/Comment.java
@@ -1,6 +1,8 @@
-package com.ourdressingtable.community.domain;
+package com.ourdressingtable.community.domain.comment;
 
+import com.ourdressingtable.community.post.domain.Post;
 import com.ourdressingtable.member.domain.Member;
+import com.ourdressingtable.util.BaseTimeEntity;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Builder;
@@ -10,12 +12,16 @@ import lombok.NoArgsConstructor;
 @Entity
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Table(name = "post_likes")
-public class PostLike {
+@Table(name = "Comments")
+public class Comment extends BaseTimeEntity {
 
     @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "post_id")
+    @Column(name = "comment_id")
     private Long id;
+
+    private String content;
+
+    private int depth;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "member_id", nullable = false)
@@ -25,10 +31,18 @@ public class PostLike {
     @JoinColumn(name = "post_id", nullable = false)
     private Post post;
 
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "parent_id", nullable = true)
+    private Comment parent;
+
     @Builder
-    public PostLike(Long id, Member member, Post post) {
-        this.id = id;
+    public Comment(String content, int depth, Member member, Post post, Comment parent) {
+        this.content = content;
+        this.depth = depth;
         this.member = member;
         this.post = post;
+        this.parent = parent;
     }
+
+
 }
