+++ backend/src/main/java/com/ourdressingtable/community/comment/repository/CommentRepository.java
@@ -1,6 +1,5 @@
-package com.ourdressingtable.community.repository;
+package com.ourdressingtable.community.domain.comment;
 
-import com.ourdressingtable.community.domain.Comment;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CommentRepository extends JpaRepository<Comment, Long> {
