+++ backend/src/main/java/com/ourdressingtable/community/post/domain/PostLike.java
@@ -1,7 +1,6 @@
-package com.ourdressingtable.community.domain;
+package com.ourdressingtable.community.post.domain;
 
 import com.ourdressingtable.member.domain.Member;
-import com.ourdressingtable.util.BaseTimeEntity;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Builder;
@@ -11,16 +10,12 @@ import lombok.NoArgsConstructor;
 @Entity
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Table(name = "Comments")
-public class Comment extends BaseTimeEntity {
+@Table(name = "post_likes")
+public class PostLike {
 
     @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "comment_id")
+    @Column(name = "post_like_id")
     private Long id;
-
-    private String content;
-
-    private int depth;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "member_id", nullable = false)
@@ -30,18 +25,10 @@ public class Comment extends BaseTimeEntity {
     @JoinColumn(name = "post_id", nullable = false)
     private Post post;
 
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "parent_id", nullable = true)
-    private Comment parent;
-
     @Builder
-    public Comment(String content, int depth, Member member, Post post, Comment parent) {
-        this.content = content;
-        this.depth = depth;
+    public PostLike(Long id, Member member, Post post) {
+        this.id = id;
         this.member = member;
         this.post = post;
-        this.parent = parent;
     }
-
-
 }
