+++ backend/src/main/java/com/ourdressingtable/community/post/repository/PostLikeRepository.java
@@ -1,6 +1,6 @@
-package com.ourdressingtable.community.repository;
+package com.ourdressingtable.community.post.repository;
 
-import com.ourdressingtable.community.domain.PostLike;
+import com.ourdressingtable.community.post.domain.PostLike;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface PostLikeRepository extends JpaRepository<PostLike, Long> {
