+++ backend/src/main/java/com/ourdressingtable/community/post/repository/PostRepository.java
@@ -1,6 +1,6 @@
-package com.ourdressingtable.community.repository;
+package com.ourdressingtable.community.post.repository;
 
-import com.ourdressingtable.community.domain.Post;
+import com.ourdressingtable.community.post.domain.Post;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface PostRepository extends JpaRepository<Post, Long> {
