+++ backend/src/main/java/com/ourdressingtable/community/service/CommunityServiceImpl.java
@@ -1,13 +1,9 @@
 package com.ourdressingtable.community.service.impl;
 
-import com.ourdressingtable.community.domain.CommunityCategory;
-import com.ourdressingtable.community.domain.Post;
-import com.ourdressingtable.community.dto.CommunityCategoryResponse;
-import com.ourdressingtable.community.dto.CreatePostRequest;
-import com.ourdressingtable.community.service.CommunityCategoryService;
+import com.ourdressingtable.community.post.dto.CreatePostRequest;
+import com.ourdressingtable.communityCategory.service.CommunityCategoryService;
 import com.ourdressingtable.community.service.CommunityService;
-import com.ourdressingtable.community.service.PostService;
-import com.ourdressingtable.member.domain.Member;
+import com.ourdressingtable.community.post.service.PostService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
