+++ backend/src/main/java/com/ourdressingtable/communityCategory/repository/CommunityCategoryRepository.java
@@ -1,6 +1,6 @@
-package com.ourdressingtable.community.repository;
+package com.ourdressingtable.communityCategory.repository;
 
-import com.ourdressingtable.community.domain.CommunityCategory;
+import com.ourdressingtable.communityCategory.domain.CommunityCategory;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CommunityCategoryRepository extends JpaRepository<CommunityCategory, Long> {
